# backend/docs/swagger.yaml
definitions:
  entity.Asset:
    properties:
      amount:
        example: 1000000
        type: integer
      asset_type:
        type: string
      code:
        example: USDC
        type: string
      distributor:
        $ref: '#/definitions/entity.Wallet'
      id:
        example: 1
        type: integer
      image:
        items:
          type: integer
        type: array
      issuer:
        $ref: '#/definitions/entity.Wallet'
      name:
        example: USD Coin
        type: string
    type: object
...
  entity.Key:
    properties:
      id:
        example: 1
        type: integer
      publicKey:
        example: GCK...
        type: string
      walletId:
        example: 1
        type: integer
      weight:
        example: 3
        type: integer
    type: object
  entity.LogTransaction:
    properties:
      amount:
        example: 100000
        type: number
      asset:
        $ref: '#/definitions/entity.Asset'
      current_main_vault:
        example: 1000
        type: number
      current_supply:
        example: 1000
        type: number
      date:
        example: "2023-08-10T14:30:00Z"
        type: string
      description:
        example: Mint Asset
        type: string
      destination_pk:
        example: GSDC...
        type: string
      log_id:
        example: 1
        type: integer
      origin_pk:
        example: GSDC...
        type: string
      transaction_type_id:
        example: 1
        type: integer
      user_id:
        example: 42
        type: integer
    type: object
...
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      token:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.UserRole:
    properties:
      id_role:
        type: string
      id_user:
        type: string
    type: object
...
  entity.Wallet:
    properties:
      funded:
        type: boolean
      id:
        example: 1
        type: integer
      key:
        $ref: '#/definitions/entity.Key'
      type:
        example: sponsor
        type: string
    type: object
...
  v1.response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  v1.userResponse:
    properties:
      user:
        $ref: '#/definitions/entity.User'
    type: object
info:
  contact: {}
paths:
...
  /log_transactions/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get all transactions logs for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LogTransaction'
      security:
      - ApiKeyAuth: []
      summary: Get transactions logs by User ID
      tags:
      - Log Transactions
...
  /user/create:
    post:
      consumes:
      - application/json
      description: Create user
      operationId: create
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.userResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create user
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: Autentication User
      operationId: autentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.userResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Autentication User
      tags:
      - user
  /user/logout:
    post:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.userResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Logout User
      tags:
      - user
  /users/edit-users-role:
    post:
      consumes:
      - application/json
      description: Edit users role
      operationId: editUsersRole
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserRole'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Edit users role
      tags:
      - user
...
swagger: "2.0"